pyproject.toml
[tool.poetry]
name = "core-tutorials"
version = "0.1.0"
description = ""
authors = ["omertairi04 <otairi010@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
djangorestframework = "^3.15.2"
psycopg = "^3.2.1"
django-filter = "^24.2"
psycopg2 = "^2.9.9"
django-split-settings = "^1.3.2"
PyYAML = "^6.0.1"
split-settings = "^1.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^7.1.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
